h24-ai-app/requirements.txt
pandas
tqdm
google-cloud-documentai
pymupdf
json
os
sys
logging
itertools
typing